+++ src/App.tsx
@@ -1,23 +1,17 @@
+import React from 'react';
 import { Route, Routes } from 'react-router-dom';
+import { UserPage } from './pages/UserPage/UserPage';
 import { LoginPage } from './pages/LoginPage/LoginPage';
-import { Navigation } from './components/Navigation/Navigation';
-import { PortfolioPage } from './pages/PortfolioPage/PortfolioPage';
 import { AdminViewPage } from './pages/AdminViewPage/AdminViewPage';
 
 function App() {
   return (
     <Routes>
       <Route path={'/'} element={<LoginPage />} />
-      <Route
-        path={'/portfolio'}
-        element={
-          <>
-            <Navigation />
-            <PortfolioPage />
-          </>
-        }
-      />
       <Route path={'/admin'} element={<AdminViewPage />} />
+      <Route path='/user/:id' element={<UserPage />} />
+      <Route path={'/user'} element={<UserPage />} />
+      <Route path={'*'} element={<LoginPage />} />
     </Routes>
   );
 }