+++ src/pages/LoginPage/LoginPage.tsx
@@ -1,8 +1,8 @@
 import './_login_page.scss';
 import { Logo } from '../../components/Logo/Logo';
 import { LoginForm } from '../../components/LoginForm/LoginForm';
-
-export const LoginPage = () => {
+type Props = {};
+export const LoginPage = (props: Props) => {
   return (
     <div className='login-page'>
       <div className='login-page__container'>