+++ CHANGELOG.md
@@ -11,4 +11,8 @@
 - [MKHH-#5] Plik json do textów statycznych, plik ze stałymi dla sass
 - [MKHH-#13] Strona logowania -- Stworzenie komponentów reużywalnych: button
 - [MKHH-#14] Strona logowania -- Projekt layoutu strony
+
+## Paweł Baraniak
+
 - [MKHH-#40] Połączenie typów Front End z Back End
+- [MKHH-#20] Widok użytkownika -- Projekt layoutu strony